+++ oap-server/server-query-plugin/query-graphql-plugin/src/main/java/org/apache/skywalking/oap/query/graphql/resolver/AlarmQuery.java
@@ -21,23 +21,39 @@ package org.apache.skywalking.oap.query.graphql.resolver;
 import com.coxautodev.graphql.tools.GraphQLQueryResolver;
 import java.io.IOException;
 import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.stream.Collectors;
 
+import org.apache.commons.lang3.StringUtils;
 import org.apache.skywalking.oap.server.core.CoreModule;
+import org.apache.skywalking.oap.server.core.analysis.IDManager;
 import org.apache.skywalking.oap.server.core.analysis.manual.searchtag.Tag;
 import org.apache.skywalking.oap.server.core.query.AlarmQueryService;
+import org.apache.skywalking.oap.server.core.query.EventQueryService;
 import org.apache.skywalking.oap.server.core.query.enumeration.Scope;
 import org.apache.skywalking.oap.server.core.query.input.Duration;
 import org.apache.skywalking.oap.server.core.query.type.AlarmTrend;
 import org.apache.skywalking.oap.server.core.query.type.Alarms;
 import org.apache.skywalking.oap.server.core.query.type.Pagination;
+import org.apache.skywalking.oap.server.core.query.type.event.Event;
+import org.apache.skywalking.oap.server.core.query.type.event.EventQueryCondition;
+import org.apache.skywalking.oap.server.core.query.type.event.EventType;
+import org.apache.skywalking.oap.server.core.query.type.event.Events;
+import org.apache.skywalking.oap.server.core.source.DefaultScopeDefine;
 import org.apache.skywalking.oap.server.library.module.ModuleManager;
+import org.apache.skywalking.oap.server.library.util.CollectionUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import static java.util.Objects.nonNull;
 
 public class AlarmQuery implements GraphQLQueryResolver {
+    private static final Logger LOGGER = LoggerFactory.getLogger(AlarmQuery.class);
 
     private final ModuleManager moduleManager;
     private AlarmQueryService queryService;
+    private EventQueryService eventQueryService;
 
     public AlarmQuery(ModuleManager moduleManager) {
         this.moduleManager = moduleManager;
@@ -48,6 +64,13 @@ public class AlarmQuery implements GraphQLQueryResolver {
             this.queryService = moduleManager.find(CoreModule.NAME).provider().getService(AlarmQueryService.class);
         }
         return queryService;
+    }
+
+    private EventQueryService getEventQueryService() {
+        if (eventQueryService == null) {
+            this.eventQueryService = moduleManager.find(CoreModule.NAME).provider().getService(EventQueryService.class);
+        }
+        return eventQueryService;
     }
 
     public AlarmTrend getAlarmTrend(final Duration duration) {
@@ -62,11 +85,166 @@ public class AlarmQuery implements GraphQLQueryResolver {
         }
         long startSecondTB = 0;
         long endSecondTB = 0;
+        EventQueryCondition condition = new EventQueryCondition();
         if (nonNull(duration)) {
             startSecondTB = duration.getStartTimeBucketInSec();
             endSecondTB = duration.getEndTimeBucketInSec();
+            condition.setTime(duration);
         }
-        return getQueryService().getAlarm(
-            scopeId, keyword, paging, startSecondTB, endSecondTB, tags);
+        condition.setType(EventType.Error);
+        condition.setName("Alarm");
+        Alarms alarms = getQueryService().getAlarm(
+                scopeId, keyword, paging, startSecondTB, endSecondTB, tags);
+        Events events = null;
+        try {
+            events = getEventQueryService().queryEvents(condition);
+        } catch (Throwable e) {
+            LOGGER.error(e.getMessage(), e);
+            return alarms;
+        }
+        return includeEvents2Alarms(alarms, events);
+    }
+
+    private Alarms includeEvents2Alarms(Alarms alarms, Events events) {
+        if (alarms.getTotal() < 1 || CollectionUtils.isEmpty(alarms.getMsgs()) || events.getTotal() < 1 || CollectionUtils.isEmpty(events.getEvents())) {
+            return alarms;
+        }
+        Map<String, List<Event>> mappingMap = events.getEvents().stream().collect(Collectors.groupingBy(Event::getServiceInSource));
+        alarms.getMsgs().forEach(a -> {
+            switch (a.getScopeId()) {
+                case DefaultScopeDefine.SERVICE :
+                    List<Event> serviceEvent = mappingMap.get(IDManager.ServiceID.analysisId(a.getId()).getName());
+                    if (CollectionUtils.isNotEmpty(serviceEvent)) {
+                        if (serviceEvent.size() == 1) {
+                            a.setEvents(serviceEvent);
+                        } else {
+                            Event event = serviceEvent.stream().filter(e -> StringUtils.equals(a.getMessage(), e.getMessage())).findFirst().orElse(null);
+                            if (Objects.nonNull(event)) {
+                                a.getEvents().set(0, event);
+                            }
+                        }
+                    }
+                    break;
+                case DefaultScopeDefine.SERVICE_RELATION :
+                    List<Event> sourceServiceEvent = mappingMap.get(IDManager.ServiceID.analysisId(a.getId()));
+                    List<Event> destServiceEvent = mappingMap.get(IDManager.ServiceID.analysisId(a.getId1()));
+                    if (CollectionUtils.isNotEmpty(sourceServiceEvent) && CollectionUtils.isNotEmpty(destServiceEvent)) {
+                        if (sourceServiceEvent.size() != 1) {
+                            Event event = sourceServiceEvent.stream().filter(e -> StringUtils.equals(a.getMessage(), e.getMessage())).findFirst().orElse(null);
+                            if (Objects.nonNull(event)) {
+                                a.getEvents().set(0, event);
+                            }
+                        } else {
+                            a.getEvents().set(0, sourceServiceEvent.get(0));
+                        }
+
+                        if (destServiceEvent.size() != 1) {
+                            Event event = destServiceEvent.stream().filter(e -> StringUtils.equals(a.getMessage(), e.getMessage())).findAny().orElse(null);
+                            if (Objects.nonNull(event)) {
+                                a.getEvents().set(1, event);
+                            }
+                        } else {
+                            a.getEvents().set(1, destServiceEvent.get(0));
+                        }
+                    }
+                    break;
+                case DefaultScopeDefine.SERVICE_INSTANCE :
+                    IDManager.ServiceInstanceID.InstanceIDDefinition instanceIDDefinition = IDManager.ServiceInstanceID.analysisId(a.getId());
+                    String serviceInstanceName = instanceIDDefinition.getName();
+                    String serviceName = IDManager.ServiceID.analysisId(instanceIDDefinition.getServiceId()).getName();
+                    List<Event> serviceInstanceEvent = mappingMap.get(serviceName);
+                    if (CollectionUtils.isNotEmpty(serviceInstanceEvent)) {
+                        if (serviceInstanceEvent.size() == 1) {
+                            a.setEvents(serviceInstanceEvent);
+                        } else {
+                            Event event = serviceInstanceEvent.stream().filter(e -> StringUtils.equals(e.getSource().getServiceInstance(), serviceInstanceName) && StringUtils.equals(e.getMessage(), a.getMessage())).findFirst().orElse(null);
+                            if (Objects.nonNull(event)) {
+                                a.getEvents().set(0, event);
+                            }
+                        }
+                    }
+                    break;
+                case DefaultScopeDefine.SERVICE_INSTANCE_RELATION :
+                    IDManager.ServiceInstanceID.InstanceIDDefinition sourceInstanceIDDefinition = IDManager.ServiceInstanceID.analysisId(a.getId());
+                    String sourceServiceInstanceName = sourceInstanceIDDefinition.getName();
+                    String sourceServiceName = IDManager.ServiceID.analysisId(sourceInstanceIDDefinition.getServiceId()).getName();
+
+                    IDManager.ServiceInstanceID.InstanceIDDefinition destInstanceIDDefinition = IDManager.ServiceInstanceID.analysisId(a.getId1());
+                    String destServiceInstanceName = destInstanceIDDefinition.getName();
+                    String destServiceName = IDManager.ServiceID.analysisId(destInstanceIDDefinition.getServiceId()).getName();
+
+                    List<Event> sourceInstanceEvent = mappingMap.get(sourceServiceName);
+                    List<Event> destInstanceEvent = mappingMap.get(destServiceName);
+                    if (CollectionUtils.isNotEmpty(sourceInstanceEvent) && CollectionUtils.isNotEmpty(destInstanceEvent)) {
+                        if (sourceInstanceEvent.size() != 1) {
+                            Event event = sourceInstanceEvent.stream().filter(e -> StringUtils.equals(e.getSource().getServiceInstance(), sourceServiceInstanceName) && StringUtils.equals(e.getMessage(), a.getMessage())).findFirst().orElse(null);
+                            if (Objects.nonNull(event)) {
+                                a.getEvents().set(0, event);
+                            }
+                        } else {
+                            a.getEvents().set(0, sourceInstanceEvent.get(0));
+                        }
+
+                        if (destInstanceEvent.size() != 1) {
+                            Event event = destInstanceEvent.stream().filter(e -> StringUtils.equals(e.getSource().getServiceInstance(), destServiceInstanceName) && StringUtils.equals(e.getMessage(), a.getMessage())).findFirst().orElse(null);
+                            if (Objects.nonNull(event)) {
+                                a.getEvents().set(1, event);
+                            }
+                        } else {
+                            a.getEvents().set(1, destInstanceEvent.get(0));
+                        }
+                    }
+                    break;
+                case DefaultScopeDefine.ENDPOINT :
+                    IDManager.EndpointID.EndpointIDDefinition endpointIDDefinition = IDManager.EndpointID.analysisId(a.getId());
+                    String endpointName = endpointIDDefinition.getEndpointName();
+                    String endpointServiceName = IDManager.ServiceID.analysisId(endpointIDDefinition.getServiceId()).getName();
+                    List<Event> serviceEndpointEvent = mappingMap.get(endpointServiceName);
+                    if (CollectionUtils.isNotEmpty(serviceEndpointEvent)) {
+                        if (serviceEndpointEvent.size() == 1) {
+                            a.setEvents(serviceEndpointEvent);
+                        } else {
+                            Event event = serviceEndpointEvent.stream().filter(e -> StringUtils.equals(e.getSource().getEndpoint(), endpointName) && StringUtils.equals(e.getMessage(), a.getMessage())).findFirst().orElse(null);
+                            if (Objects.nonNull(event)) {
+                                a.getEvents().set(0, event);
+                            }
+                        }
+                    }
+                    break;
+                case DefaultScopeDefine.ENDPOINT_RELATION :
+                    IDManager.EndpointID.EndpointIDDefinition sourceEndpointIDDefinition = IDManager.EndpointID.analysisId(a.getId());
+                    String sourceEndpointName = sourceEndpointIDDefinition.getEndpointName();
+                    String sourceEndpointServiceName = IDManager.ServiceID.analysisId(sourceEndpointIDDefinition.getServiceId()).getName();
+
+                    IDManager.EndpointID.EndpointIDDefinition destEndpointIDDefinition = IDManager.EndpointID.analysisId(a.getId1());
+                    String destEndpointName = destEndpointIDDefinition.getEndpointName();
+                    String destEndpointServiceName = IDManager.ServiceID.analysisId(destEndpointIDDefinition.getServiceId()).getName();
+
+                    List<Event> sourceEndpointEvent = mappingMap.get(sourceEndpointServiceName);
+                    List<Event> destEndpointEvent = mappingMap.get(destEndpointServiceName);
+
+                    if (CollectionUtils.isNotEmpty(sourceEndpointEvent) && CollectionUtils.isNotEmpty(destEndpointEvent)) {
+                        if (sourceEndpointEvent.size() != 1) {
+                            Event event = sourceEndpointEvent.stream().filter(e -> StringUtils.equals(e.getSource().getEndpoint(), sourceEndpointName) && (StringUtils.equals(e.getMessage(), a.getMessage()))).findFirst().orElse(null);
+                            if (Objects.nonNull(event)) {
+                                a.getEvents().set(0, event);
+                            }
+                        } else {
+                            a.getEvents().set(0, sourceEndpointEvent.get(0));
+                        }
+
+                        if (destEndpointEvent.size() != 1) {
+                            Event event = destEndpointEvent.stream().filter(e -> StringUtils.equals(e.getSource().getEndpoint(), destEndpointName) && (StringUtils.equals(e.getMessage(), a.getMessage()))).findFirst().orElse(null);
+                            if (Objects.nonNull(event)) {
+                                a.getEvents().set(1, event);
+                            }
+                        } else {
+                            a.getEvents().set(1, destEndpointEvent.get(0));
+                        }
+                    }
+                    break;
+            }
+        });
+        return alarms;
     }
 }